Fix Errors and Deprecations for Http Project
