+++ http/src/main/scala/org/http4s/blaze/http/package.scala
@@ -11,21 +11,4 @@ package object http {
   // The basic type that represents a HTTP service
   type HttpService = HttpRequest => Future[RouteAction]
 
-  /** Standard HTTP request
-    *
-    * @param method HTTP request method
-    * @param uri request uri
-    * @param headers request headers
-    * @param body function which returns the next chunk of the request body. Termination is
-    *             signaled by an __empty__ `ByteBuffer` as determined by `ByteBuffer.hasRemaining()`.
-    */
-  case class HttpRequest(method: Method, uri: Uri, majorVersion: Int, minorVersion: Int, headers: Headers, body: MessageBody)
-
-  /** The prelude of a standard HTTP response
-    *
-    * @param code Response status code
-    * @param status Response message. This has no meaning for the protocol, its purely for human enjoyment.
-    * @param headers Response headers.
-    */
-  case class HttpResponsePrelude(code: Int, status: String, headers: Headers)
 }
